Canadian COVID-19 case rates

Data set for all Canadian provinces and territories containing COVID-19 case rates (COVID-19 cases per 100,000 people) derived from COVID-19 case counts as reported by the COVID-19 Canada Open Data Working Group (CCODWG). Data is available both through the archived ccodwg/Covid19Canada GitHub repository and the newer ccodwg/CovidTimelineCanada GitHub repository, which also reports vaccine-related signals.

Usage

can_prov_cases

Format

An object of class epi_df (inherits from tbl_df, tbl, data.frame) with 65299 rows and 4 columns.

Source

This object contains a modified part of the COVID-19 Canada Open Data Working Group's Covid19Canada data repository (archived). This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by the COVID-19 Canada Open Data Working Group.

Details

This dataset contains versioned data covering the period from April 2020 to December 2021 and is used in the epipredict slide vignette.