Daily COVID-19 case and death rates from all states in archive format

Data set containing COVID-19 case and death rates (counts per 100000 population) as reported by the Delphi API, based on reports made available by the Center for Systems Science and Engineering at Johns Hopkins University. This example data ranges from March 1, 2020 to November 30, 2021, issued monthly on the first day of each month from September 1, 2020 to December 1, 2021. It includes all US states, Washington DC, Guam, Puerto Rico, and the Virgin Islands.

Usage

case_death_rate_archive

Format

An epiprocess::epi_archive. The DT attribute contains the data formatted as a data.table::data.table (object of class c("data.table", "data.frame")) with 72086 rows and 7 columns.

Source

This object contains a modified part of the COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University as republished in the COVIDcast Epidata API. This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by Johns Hopkins University on behalf of its Center for Systems Science in Engineering. Copyright Johns Hopkins University 2020.

Modifications:

From the COVIDcast Epidata API: The signals case_rate and death_rate are taken directly from the JHU CSSE GitHub repo without changes, served through the Delphi API.

From the COVIDcast Epidata API: Averaged signals were computed from the original JHU-CSSE data by calculating moving averages of the preceding 7 days, so the signal for June 7 is the average of the underlying data for June 1 through 7, inclusive.

Furthermore, the data has been limited to a specific time range, the signal names slightly altered, and formatted into an epi_archive.

Data dictionary

The data in the epi_archive$DT attribute has columns:

geo_value

the geographic value associated with each row of measurements.

time_value

the time value associated with each row of measurements.

version

the time value specifying the version for each row of measurements.

case_rate

Number of new confirmed cases due to COVID-19 per 100,000 population, daily

case_rate_7d_av

7-day average signal of number of new confirmed cases due to COVID-19 per 100,000 population, daily

death_rate

Number of new confirmed deaths due to COVID-19 per 100,000 population, daily

death_rate_7d_av

7-day average signal of number of new confirmed deaths due to COVID-19 per 100,000 population, daily