Subset of COVID-19 incident cases and deaths from 5 states

Data set for 5 states containing COVID-19 Incident Cases and Deaths as reported by JHU-CSSE and downloaded from the CMU Delphi COVIDcast Epidata API. This example data is a snapshot as of March 20, 2024, and ranges from June 4, 2021 to December 31, 2021. It is limited to California, Florida, Texas, New Jersey, and New York.

Usage

counts_subset

Format

An epiprocess::epi_df (object of class c("epi_df", "tbl_df", "tbl", "data.frame")) with 1055 rows and 4 columns.

Source

This object contains a modified part of the COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University as republished in the COVIDcast Epidata API. This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering. Copyright Johns Hopkins University 2020.

Modifications:

From the COVIDcast Epidata API: The signals are taken directly from the JHU CSSE COVID-19 GitHub repository without changes.

Furthermore, the data has been limited to a very small number of rows, the signal names slightly altered, and formatted into an epi_df.

Data dictionary

The data has columns:

geo_value

the geographic value associated with each row of measurements.

time_value

the time value associated with each row of measurements.

cases

Number of new confirmed COVID-19 cases, daily

deaths

Number of new confirmed COVID-19 deaths, daily