Subset of COVID-19 cumulative case counts from 4 states

Data set for 4 states containing COVID-19 Cumulative Cases as reported by JHU-CSSE and downloaded from the CMU Delphi COVIDcast Epidata API. This example data is a snapshot as of March 20, 2024, and ranges from March 1, 2020 to January 31, 2022. It is limited to California, Florida, New York, and Texas.

It is used in the epiprocess "Getting Started" vignette.

Usage

covid_confirmed_cumulative_num

Format

A tibble::tibble (object of class c("tbl_df", "tbl", "data.frame")) with 2808 rows and 15 columns.

Source

This object contains a modified part of the COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University as republished in the COVIDcast Epidata API. This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering. Copyright Johns Hopkins University 2020.

Modifications:

From the COVIDcast Epidata API: These signals are taken directly from the JHU CSSE COVID-19 GitHub repository without changes.

Furthermore, the data has been limited to a very small number of rows, formatted into an epi_df, and the signal names slightly altered.

Data dictionary

The data has columns:

geo_value

the geographic value associated with each row of measurements.

signal

name of metric, derived from upstream data.

source

name of upstream data source.

geo_type

spatial resolution of the signal.

time_type

temporal resolution of the signal.

time_value

the time value associated with each row of measurements.

issue

time unit (e.g., date) when the signal data were published.

lag

time delta (e.g. days) between when the underlying events happened and when the data were published.

missing_value

an integer code that is zero when the value field is present and non-zero when the data is missing (see missing codes).

missing_stderr

an integer code that is zero when the stderr field is present and non-zero when the data is missing (see missing codes).

missing_sample_size

an integer code that is zero when the sample_size field is present and non-zero when the data is missing (see missing codes).

value

cumulative number of confirmed COVID-19 cases, derived from the underlying data source.

stderr

approximate standard error of the statistic with respect to its sampling distribution, NA when not applicable.

sample_size

number of “data points” used in computing the statistic, NA when not applicable.