epidatasets: Epidemiological Data for Delphi Tooling Examples

This package contains data sets used to compile vignettes and other documentation in Delphi R Packages. The goal is to avoid calls to the Delphi Epidata API, and to deposit some examples here for easy offline use.

See also

Author

Maintainer: Nat DeFries ndefries@andrew.cmu.edu

Authors:

Daniel J. McDonald daniel@stat.ubc.ca

Other contributors:

Johns Hopkins University Center for Systems Science and Engineering (Owner of COVID-19 cases and deaths data from the COVID-19 Data Repository) [data contributor]

Johns Hopkins University (Copyright holder of COVID-19 cases and deaths data from the COVID-19 Data Repository) [copyright holder]

Carnegie Mellon University Delphi Group (Owner of masking, social-distancing, and CLI data from the COVID-19 Trends and Impacts Survey. Owner of claims-based CLI data from the Delphi Epidata API) [data contributor]

The COVID-19 Canada Open Data Working Group (Owner of Canadian COVID-19 cases rates from the Covid19Canada data repository) [data contributor]

Statistics Canada (Owner of Canadian graduate employment income data from the Statistics Canada website) [data contributor]

Google (Collaborator on CLI data from the Google symptom surveys) [data contributor]