Subset of Statistics Canada median employment income for postsecondary graduates

Data set for all territories (aggregated) and all 10 provinces containing yearly income data for postsecondary graduates as reported by Statistics Canada, downloaded from the Statistics Canada website at www.statcan.gc.ca. This example data is a snapshot as of September 18, 2024, and ranges from 2010 to 2017 (yearly).

Usage

grad_employ_subset

Format

An epiprocess::epi_df (object of class c("epi_df", "tbl_df", "tbl", "data.frame")) with 1445 rows and 7 columns.

Source

This object contains modified data adapted from Statistics Canada, Table 37-10-0115-01 Characteristics and median employment income of longitudinal cohorts of postsecondary graduates two and five years after graduation, by educational qualification and field of study (primary groupings). This does not constitute an endorsement by Statistics Canada of this product.

The data is licensed under the terms of the Statistics Canada Open License.

Modifications:

Only provincial and territorial regions are kept.

Only age group, field of study, and educational qualification are kept as covariates. For the remaining covariates, we keep aggregated values and drop the level-specific rows.

No modifications were made to the time range of the data.

Data dictionary

The data has columns:

geo_value

The province in Canada associated with each row of measurements.

time_value

The time value, a year integer in YYYY format

edu_qual

The education qualification

age_group

The age group; either 15 to 34 or 35 to 64

num_graduates

The number of graduates for the given row of characteristics

med_income_2y

The median employment income two years after graduation

med_income_5y

The median employment income five years after graduation